Refactoring the XpertResult
According to the pervious comment, now the XpertResult on the constructor builds a vector of XpertRequestResult that contains a drugTreatment, an error message and the XpertRequestData. An XpertQueryToCoreExtractor that extends QueryToCoreExtractor had been implemented, it returns a drug treatment only if there is only one drug matching. After the XpertResult construction, the XpertQueryData pointer is invalidated because it is moved into the constructor.