Maintenance
Removing white spaces from xml
Removing localComputation optional
Adjusting ClinicalData with clinicalDataEntries and xsd validation
Exporting main in a separate class
Adding the possibility to set a "date of execution" in the ModelSelector to avoir using DateTime::now and to allow testing to be idempotent 
Utils testing (double decimal doubleToStr and langToStr).
Handling Translatable string in model selector for validation, name and description.