Header, drug intro and copyright html export
Lost a lot of time trying to use a css framework but it is incompatible with the pdf exporter. It is incompatible because they use flex box but the exporter is too old to understand it. Consequently, I learnt new way to align things with classic tables. Made the header, drug intro and copyright.